feat(email): restructure email module and implement Kafka consumer integration
