+++ apps/email/src/consumer/services/email.service.ts
@@ -1,14 +1,17 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { KafkaConsumer } from 'libs/core/kafka/kafka.interface';
 
-import { EmailService } from './email.service';
-import { EmailPayload, EmailToken } from '../../core/email.interface';
+import {
+  EmailInterface,
+  EmailPayload,
+  EmailToken,
+} from '../../email/core/email.interface';
 
 @Injectable()
 export class EmailKafkaConsumer implements KafkaConsumer<EmailPayload> {
   constructor(
     @Inject(EmailToken)
-    private readonly emailService: EmailService,
+    private readonly emailService: EmailInterface,
   ) {}
 
   async handleEvent(payload: EmailPayload): Promise<void> {
