+++ apps/email/src/email/core/email.interface.ts
