+++ apps/email/src/email/email.module.ts
@@ -1,0 +1,25 @@
+import { Module } from '@nestjs/common';
+import { HttpClientModule } from 'libs/infrastructure/http/http-client.module';
+
+import { EmailTransportToken } from './core/email-transport.interface';
+import { EmailToken } from './core/email.interface';
+import { NodemailerService } from './infrastructure/providers/nodemailer.provider';
+import { EmailService } from './infrastructure/services/email.service';
+import { EmailController } from './interface/email.controller';
+
+@Module({
+  imports: [HttpClientModule],
+  providers: [
+    {
+      provide: EmailTransportToken,
+      useClass: NodemailerService,
+    },
+    {
+      provide: EmailToken,
+      useClass: EmailService,
+    },
+  ],
+  controllers: [EmailController],
+  exports: [EmailToken],
+})
+export class EmailModule {}
