--- libs/infrastructure/kafka/kafka.consumer.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-
-import { kafkaPublisherProvider } from './kafka.publisher';
-
-@Module({
-  providers: [kafkaPublisherProvider],
-  exports: [kafkaPublisherProvider],
-})
-export class KafkaPublisherModule {}
